+++ package.json
@@ -4,7 +4,9 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "jest --runInBand",
+    "test-arranger-acl": "jest --testPathPattern='src/schemas/Arranger/test/metadataAccessControl.test.ts'",
+    "test-base": "jest --runInBand --testPathIgnorePatterns='src/schemas/Arranger/test/metadataAccessControl.test.ts'",
+    "test": "npm run test-base && npm run test-arranger-acl",
     "dev": "NODE_PATH=./src ts-node-dev ./src/index.ts -p ./tsconfig.json",
     "build": "rm -rf dist && tsc -p ./tsconfig.build.json && cp -r ./src/resources ./dist",
     "start::prod": "NODE_PATH=./dist node dist/index.js",
