+++ src/schemas/Arranger/test/index.test.ts
@@ -28,18 +28,7 @@ import { Client } from '@elastic/elasticsearch';
 import getArrangerGqlSchema from '..';
 import { ARRANGER_FILE_CENTRIC_INDEX, ARRANGER_PROJECT_ID } from 'config';
 import metadata from 'resources/arranger_es_metadata.json';
-import { exec } from 'child_process';
-import { promisify } from 'util';
-import { EsFileCentricDocument, FILE_RELEASE_STAGE } from 'utils/commonTypes/EsFileCentricDocument';
-import {
-  getAllIndexedDocuments,
-  MOCK_API_KEYS,
-  MOCK_API_KEY_SCOPES,
-} from 'routes/file-storage-api/handlers/test/utils';
-import { aggregateAllObjectIds, fileDocumentStream, reduceToFileHits } from './utils';
 import _ from 'lodash';
-
-const asyncExec = promisify(exec);
 
 describe('Arranger schema', () => {
   const mockMapping = {
@@ -81,9 +70,10 @@ describe('Arranger schema', () => {
       body: mockMapping,
     });
   }, 120000);
-  afterAll(async () => {
+  afterAll( async (done) => {
     await esContainer.stop();
-  }, 120000);
+    done()
+  });
 
   describe('initArrangerMetadata', () => {
     it('indices must exist after run', async () => {
@@ -162,231 +152,6 @@ describe('Arranger schema', () => {
       } catch (err) {
         throw err;
       }
-    });
-  });
-
-  describe('metadata access control', () => {
-    let allIndexedDocuments: { [objectId: string]: EsFileCentricDocument } = {};
-
-    beforeAll(async () => {
-      const { stdout, stderr } = await asyncExec(`ES_HOST=${esHost} npm run releaseStageEsInit`);
-      if (stderr.length) {
-        throw stderr;
-      }
-      console.log("releaseStageEsInit stdout: ", stdout);
-      await new Promise(resolve => {
-        setTimeout(() => {
-          resolve();
-        }, 30000);
-      });
-
-      allIndexedDocuments = (await getAllIndexedDocuments(esClient)).reduce(
-        (acc, doc) => {
-          acc[doc.object_id] = doc;
-          return acc;
-        },
-        {} as typeof allIndexedDocuments,
-      );
-    });
-
-    describe('hits query', () => {
-      it('returns unique entities', async () => {
-        const responseStream = fileDocumentStream({
-          esClient,
-          apiKey: MOCK_API_KEYS.PUBLIC,
-        });
-        const allRetrievedEntities = await reduceToFileHits(responseStream);
-        expect(
-          _(allRetrievedEntities)
-            .uniqBy(e => e.node.object_id)
-            .size(),
-        ).toBe(allRetrievedEntities.length);
-      }, 240000);
-
-      it('returns and all the right data for public users', async () => {
-        const responseStream = fileDocumentStream({
-          esClient,
-          apiKey: MOCK_API_KEYS.PUBLIC,
-        });
-        const allEntityIdsFromApi = (await reduceToFileHits(responseStream)).map(
-          e => e.node.object_id,
-        );
-        const equivalentIndexedDocuments = allEntityIdsFromApi.map(
-          id => allIndexedDocuments[id || ''],
-        );
-        const allDocumentsThatQualify = Object.values(allIndexedDocuments).filter(
-          doc => doc.release_stage === FILE_RELEASE_STAGE.PUBLIC,
-        );
-        expect(equivalentIndexedDocuments.length).toBe(allDocumentsThatQualify.length);
-        expect(allDocumentsThatQualify.every(doc => equivalentIndexedDocuments.includes(doc))).toBe(
-          true,
-        );
-        expect(equivalentIndexedDocuments.every(doc => allDocumentsThatQualify.includes(doc))).toBe(
-          true,
-        );
-      }, 240000);
-
-      it('returns all data for DCC', async () => {
-        const responseStream = fileDocumentStream({
-          esClient,
-          apiKey: MOCK_API_KEYS.DCC,
-        });
-        const allRetrievedEntities = await reduceToFileHits(responseStream);
-        expect(allRetrievedEntities.length).toBe(Object.entries(allIndexedDocuments).length);
-      }, 240000);
-
-      it('returns and all the right data for program members', async () => {
-        const apiKey = MOCK_API_KEYS.FULL_PROGRAM_MEMBER;
-        const userScopes = MOCK_API_KEY_SCOPES[apiKey];
-        const responseStream = fileDocumentStream({ esClient, apiKey: apiKey });
-        const allRetrievedEntities = await reduceToFileHits(responseStream);
-        const equivalentIndexedDocuments = allRetrievedEntities.map(
-          retrievedObject => allIndexedDocuments[retrievedObject.node.object_id || ''],
-        );
-        const validators: ((doc: EsFileCentricDocument) => boolean)[] = [
-          ({ release_stage }) => release_stage === FILE_RELEASE_STAGE.PUBLIC,
-          ({ release_stage }) => release_stage === FILE_RELEASE_STAGE.PUBLIC_QUEUE,
-          ({ release_stage }) => release_stage === FILE_RELEASE_STAGE.FULL_PROGRAMS,
-          ({ release_stage }) => release_stage === FILE_RELEASE_STAGE.ASSOCIATE_PROGRAMS,
-          ({ study_id, release_stage }) =>
-            release_stage === FILE_RELEASE_STAGE.OWN_PROGRAM &&
-            userScopes.some(scope => scope.includes(study_id)),
-        ];
-        const allDocumentsThatQualify = Object.values(allIndexedDocuments).filter(doc =>
-          validators.some(validate => validate(doc)),
-        );
-        expect(equivalentIndexedDocuments.length).toBe(allDocumentsThatQualify.length);
-        expect(equivalentIndexedDocuments.every(doc => allDocumentsThatQualify.includes(doc))).toBe(
-          true,
-        );
-        expect(allDocumentsThatQualify.every(doc => equivalentIndexedDocuments.includes(doc))).toBe(
-          true,
-        );
-      }, 240000);
-
-      it('returns and all the right data for associate program members', async () => {
-        const apiKey = MOCK_API_KEYS.ASSOCIATE_PROGRAM_MEMBER;
-        const userScopes = MOCK_API_KEY_SCOPES[apiKey];
-        const responseStream = fileDocumentStream({ esClient, apiKey: apiKey });
-        const allRetrievedEntities = await reduceToFileHits(responseStream);
-        const equivalentIndexedDocuments = allRetrievedEntities.map(
-          retrievedObject => allIndexedDocuments[retrievedObject.node.object_id || ''],
-        );
-        const validators: ((doc: EsFileCentricDocument) => boolean)[] = [
-          ({ release_stage }) => release_stage === FILE_RELEASE_STAGE.PUBLIC,
-          ({ release_stage }) => release_stage === FILE_RELEASE_STAGE.PUBLIC_QUEUE,
-          ({ release_stage }) => release_stage === FILE_RELEASE_STAGE.ASSOCIATE_PROGRAMS,
-          ({ study_id, release_stage }) =>
-            release_stage === FILE_RELEASE_STAGE.FULL_PROGRAMS &&
-            userScopes.some(scope => scope.includes(study_id)),
-          ({ study_id, release_stage }) =>
-            release_stage === FILE_RELEASE_STAGE.OWN_PROGRAM &&
-            userScopes.some(scope => scope.includes(study_id)),
-        ];
-        const allDocumentsThatQualify = Object.values(allIndexedDocuments).filter(doc =>
-          validators.some(validate => validate(doc)),
-        );
-        expect(equivalentIndexedDocuments.length).toBe(allDocumentsThatQualify.length);
-        expect(equivalentIndexedDocuments.every(doc => allDocumentsThatQualify.includes(doc))).toBe(
-          true,
-        );
-        expect(allDocumentsThatQualify.every(doc => equivalentIndexedDocuments.includes(doc))).toBe(
-          true,
-        );
-      }, 240000);
-    });
-
-    describe('aggregation query', () => {
-
-      it('returns and all the right data for public users', async () => {
-        const aggregationResult = await aggregateAllObjectIds({
-          apiKey: MOCK_API_KEYS.PUBLIC,
-          esClient,
-        })
-        const allEntityIdsFromApi = aggregationResult?.file.aggregations.object_id.buckets.map(({key}) => key) || []
-        const equivalentIndexedDocuments = allEntityIdsFromApi.map(
-          id => allIndexedDocuments[id || ''],
-        );
-        const allDocumentsThatQualify = Object.values(allIndexedDocuments).filter(
-          doc => doc.release_stage === FILE_RELEASE_STAGE.PUBLIC,
-        );
-        expect(equivalentIndexedDocuments.length).toBe(allDocumentsThatQualify.length);
-        expect(allDocumentsThatQualify.every(doc => equivalentIndexedDocuments.includes(doc))).toBe(
-          true,
-        );
-        expect(equivalentIndexedDocuments.every(doc => allDocumentsThatQualify.includes(doc))).toBe(
-          true,
-        );
-      }, 240000);
-
-      it('returns all data for DCC', async () => {
-        const aggregationResult = await aggregateAllObjectIds({
-          apiKey: MOCK_API_KEYS.DCC,
-          esClient,
-        })
-        const allEntityIdsFromApi = aggregationResult?.file.aggregations.object_id.buckets.map(({key}) => key) || []
-        expect(allEntityIdsFromApi.length).toBe(Object.entries(allIndexedDocuments).length);
-      }, 240000);
-
-      it('returns and all the right data for program members', async () => {
-        const apiKey = MOCK_API_KEYS.FULL_PROGRAM_MEMBER;
-        const userScopes = MOCK_API_KEY_SCOPES[apiKey];
-        const aggregationResult = await aggregateAllObjectIds({ apiKey, esClient })
-        const allObjectIdsFromApi = aggregationResult?.file.aggregations.object_id.buckets.map(({key}) => key) || []
-        const equivalentIndexedDocuments = allObjectIdsFromApi.map(
-          objectId => allIndexedDocuments[objectId],
-        );
-        const validators: ((doc: EsFileCentricDocument) => boolean)[] = [
-          ({ release_stage }) => release_stage === FILE_RELEASE_STAGE.PUBLIC,
-          ({ release_stage }) => release_stage === FILE_RELEASE_STAGE.PUBLIC_QUEUE,
-          ({ release_stage }) => release_stage === FILE_RELEASE_STAGE.FULL_PROGRAMS,
-          ({ release_stage }) => release_stage === FILE_RELEASE_STAGE.ASSOCIATE_PROGRAMS,
-          ({ study_id, release_stage }) =>
-            release_stage === FILE_RELEASE_STAGE.OWN_PROGRAM &&
-            userScopes.some(scope => scope.includes(study_id)),
-        ];
-        const allDocumentsThatQualify = Object.values(allIndexedDocuments).filter(doc =>
-          validators.some(validate => validate(doc)),
-        );
-        expect(equivalentIndexedDocuments.length).toBe(allDocumentsThatQualify.length);
-        expect(equivalentIndexedDocuments.every(doc => allDocumentsThatQualify.includes(doc))).toBe(
-          true,
-        );
-        expect(allDocumentsThatQualify.every(doc => equivalentIndexedDocuments.includes(doc))).toBe(
-          true,
-        );
-      }, 240000);
-
-      it('returns and all the right data for associate program members', async () => {
-        const apiKey = MOCK_API_KEYS.ASSOCIATE_PROGRAM_MEMBER;
-        const userScopes = MOCK_API_KEY_SCOPES[apiKey];
-        const aggregationResult = await aggregateAllObjectIds({ apiKey, esClient })
-        const allObjectIdsFromApi = aggregationResult?.file.aggregations.object_id.buckets.map(({key}) => key) || []
-        const equivalentIndexedDocuments = allObjectIdsFromApi.map(
-          objectId => allIndexedDocuments[objectId],
-        );
-        const validators: ((doc: EsFileCentricDocument) => boolean)[] = [
-          ({ release_stage }) => release_stage === FILE_RELEASE_STAGE.PUBLIC,
-          ({ release_stage }) => release_stage === FILE_RELEASE_STAGE.PUBLIC_QUEUE,
-          ({ release_stage }) => release_stage === FILE_RELEASE_STAGE.ASSOCIATE_PROGRAMS,
-          ({ study_id, release_stage }) =>
-            release_stage === FILE_RELEASE_STAGE.FULL_PROGRAMS &&
-            userScopes.some(scope => scope.includes(study_id)),
-          ({ study_id, release_stage }) =>
-            release_stage === FILE_RELEASE_STAGE.OWN_PROGRAM &&
-            userScopes.some(scope => scope.includes(study_id)),
-        ];
-        const allDocumentsThatQualify = Object.values(allIndexedDocuments).filter(doc =>
-          validators.some(validate => validate(doc)),
-        );
-        expect(equivalentIndexedDocuments.length).toBe(allDocumentsThatQualify.length);
-        expect(equivalentIndexedDocuments.every(doc => allDocumentsThatQualify.includes(doc))).toBe(
-          true,
-        );
-        expect(allDocumentsThatQualify.every(doc => equivalentIndexedDocuments.includes(doc))).toBe(
-          true,
-        );
-      }, 240000);
     });
   });
 });
